refactor(web): unify number/price/latency formatters into lib/format
Consolidate the per-feature display formatters (dashboard formatInt/
formatRatio, models formatPrice/formatLatency/formatSuccessRate) into a
single web/src/lib/format.ts vocabulary and add formatCompactNumber/
formatTokens for large-value abbreviation. The dashboard barrel re-exports
the shared functions to preserve its import path.

Closes #594
